Task0

+++ Tyuiu.NazarovaKH.Sprint3.Task0.V1.Test/DataServiceTest.cs
@@ -10,7 +10,7 @@ namespace Tyuiu.NazarovaKH.Sprint3.Task0.V1.Test
         {
             DataService ds = new DataService();
 
-            double value = 4;
+            int value = 4;
             int startValue = 1;
             int stopValue = 15;
